Fix integer overflow and uint8 wrap-around in screening / index math
* vhf/fill_nr_s8.c: ij0 = i0*(i0+1)/2 + j0 was computed in int and
  overflows once i0 reaches ~65000. Promote with a (size_t) cast on
  the first factor.

* vhf/nr_sr_vhf.c: nblock and the derived nblock2/nblock3 were
  uint32_t, and blk_id was a plain int. nblock*nblock*nblock
  overflows uint32_t past nblock ~ 1626, and "int blk_id < nblock3"
  overflows int past nblock ~ 1290. Promote nblock/nblock2/nblock3
  and blk_id/r to size_t.

* mcscf/fci_rdm.c: tril_particle_symm computed
  blk = MIN(((int)(48/norb))*norb, nnorb). For norb > 48 this
  collapses to blk = 0, after which "for (m=0; m<nnorb-blk; m+=blk)"
  is an infinite loop. Use MAX(blk_units, 1) so blk stays at norb at
  minimum.

* gto/grid_ao_drv.c: screen_index used (uint8_t)(si + 1) without an
  upper bound. For si >= 255 (very tight AOs, large -arr) the cast
  wraps mod 256 and silently demotes a maximally-significant grid
  point to "screened out" (0) or to small significance. Add a
  saturating clamp at 255. Behavior for si <= 254 is unchanged.

+++ pyscf/lib/vhf/fill_nr_s8.c
@@ -92,7 +92,7 @@ static void store_ij(int (*intor)(), double *eri, double *buf, int ish, int jsh,
         for (i0 = ao_loc[ish], i = 0; i < di; i++, i0++) {
         for (j0 = ao_loc[jsh], j = 0; j < dj; j++, j0++) {
                 if (i0 >= j0) {
-                        ij0 = i0*(i0+1)/2 + j0;
+                        ij0 = (size_t)i0*(i0+1)/2 + j0;
                         peri = eri + ij0*(ij0+1)/2;
                         pbuf = buf + nao2 * (i*dj+j);
                         for (kl = 0, k = 0; k < i0; k++) {
